보안코드 부분을 컴포넌트로 분리
